+++ Fun.Htmx/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add more overloads for hxGetComponent, hxPostComponent, hxGetCustomElement, hxPostCustomElement
+
 ## [3.2.7] - 2023-11-25
 
 - Add hxGetComponent, hxPostComponent, hxGetCustomElement, hxPostCustomElement